pyproject.toml
[project]
name = "van-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Lachie Stewart", email = "lachiestewart2003@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bleak>=3.0.1",
    "paho-mqtt>=2.1.0",
    "pycryptodome>=3.23.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
    "ty>=0.0.27",
]

[project.scripts]
van = "van_assistant.main:main"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    "G004", # Ignore f-string
    "D100", # Ignore missing docstring in public module
    "D104", # Ignore missing docstring in public package
    "PLR2004", # Ignore magic values
]

[tool.pytest]
asyncio_mode = "auto"


[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"
